+++ Deconstruct/Punctuation.cs
@@ -8,16 +8,6 @@ using System.Threading.Tasks;
 
 namespace PLGL.Deconstruct
 {
-    /// <summary>
-    /// This class answers the question: "What do we do with punctuation marks?"
-    /// 
-    /// Here are several examples:
-    ///     1. For commas, the language author may wish to leave it there. This mark still needs to be added,
-    ///        otherwise it would get included into the word-seed.
-    ///     2. For question marks, the language author may wish to turn it into a word (in Japanese, the particle "ka" is used).
-    ///        If making use of non-punctuated particles (Japanese's "wa" subject particle), a flag addition is recommended.
-    ///     3. 
-    /// </summary>
     public class Punctuation
     {
         public SortedDictionary<string, Func<WordInfo, string>> Marks { get; set; }
@@ -25,7 +15,7 @@ namespace PLGL.Deconstruct
 
         public void Process(LanguageGenerator lg, WordInfo word, string filterName)
         {
-            if (word.Filter.Name.ToUpper() == filterName)
+            if (word.Filter.Name.ToUpper() == filterName && word.IsProcessed == false)
             {
                 foreach (string s in Marks.Keys)
                 {